feat: move image picker to component
