+++ frontend/src/components/ImagePicker/ImagePicker.css
@@ -1,8 +1,3 @@
-body {
-  background-color: rgb(31, 31, 31);
-  color: white;
-}
-
 .imageContainer {
   position: relative;
   display: inline-block;
